Feat[prefs]: move Vulkan driver settings onto video category
