+++ app_pojavlauncher/src/main/res/values/strings.xml
@@ -487,11 +487,9 @@
 
     <string name="driver_default_name">Built-in driver</string>
     <string name="driver_default_extra_name">MojoLauncher</string>
-    <string name="driver_config_title">AdrenoTools Drivers</string>
+    <string name="driver_config_title">AdrenoTools Driver Manager</string>
     <string name="driver_config_subtitle">Manage installed AdrenoTools Vulkan drivers and install new</string>
-    <string name="driver_config_import">Import driver</string>
-    <string name="driver_config_delete">Remove driver</string>
+    <string name="driver_config_import">Add</string>
+    <string name="driver_config_delete">Remove</string>
     <string name="driver_config_import_failed">Failed to add the selected driver!</string>
-    <string name="preference_category_vulkan_title">Vulkan Driver Settings</string>
-    <string name="preference_category_vulkan_description">Manage custom Vulkan driver settings</string>
 </resources>
