+++ app_pojavlauncher/src/main/res/xml/pref_video.xml
@@ -12,6 +12,12 @@
             android:icon="@drawable/ic_px_image_renderer"
             app2:useSimpleSummaryProvider="true"/>
 
+        <Preference
+            android:key="manageDrivers"
+            android:persistent="false"
+            android:summary="@string/driver_config_subtitle"
+            android:title="@string/driver_config_title"/>
+
         <androidx.preference.SwitchPreferenceCompat
             android:defaultValue="false"
             android:key="ignoreNotch"
@@ -56,6 +62,11 @@
             android:summary="@string/preference_use_angle_description"
             android:title="@string/preference_use_angle_title"
             />
+        <SwitchPreference
+            android:defaultValue="false"
+            android:key="zinkPreferSystemDriver"
+            android:summary="@string/preference_vulkan_driver_system_description"
+            android:title="@string/preference_vulkan_driver_system_title"/>
         <androidx.preference.SwitchPreferenceCompat
             android:defaultValue="true"
             android:key="vsync_in_zink"
