--- app_pojavlauncher/src/main/res/xml/pref_vulkan.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app2="http://schemas.android.com/apk/res-auto">
-
-    <net.kdt.pojavlaunch.prefs.BackButtonPreference/>
-
-    <PreferenceCategory android:title="@string/driver_config_title"></PreferenceCategory>
-    <Preference
-        android:key="manageDrivers"
-        android:persistent="false"
-        android:summary="@string/driver_config_subtitle"
-        android:title="@string/driver_config_title"/>
-    <SwitchPreference
-        android:defaultValue="false"
-        android:key="zinkPreferSystemDriver"
-        android:summary="@string/preference_vulkan_driver_system_description"
-        android:title="@string/preference_vulkan_driver_system_title"/>
-
-</PreferenceScreen>
